+++ composeApp/src/commonMain/kotlin/mende273/quoteskmp/domain/repository/local/LocalRepository.kt
@@ -6,7 +6,7 @@ import mende273.quoteskmp.domain.model.Quote
 interface LocalRepository {
 
     suspend fun getAllFavouriteQuotes(): Flow<List<Quote>>
-    suspend fun getFavouriteQuote(id: Int): Flow<Quote?>
+    suspend fun getFavouriteQuote(id: Long): Flow<Quote?>
     suspend fun addFavouriteQuote(quote: Quote)
     suspend fun removeFavouriteQuote(quote: Quote)
 }
